% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile.mmre}
\alias{compile.mmre}
\title{Compiling TMB C++ templates}
\usage{
compile.mmre()
}
\description{
Compiles the  TMB templates into a shared object
file. Must be done a single time following installation or
updating of the package.
}
