% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{coords2spdf}
\alias{coords2spdf}
\title{Function to make a SpatialPolygonsDataFrame from a matrix of points}
\usage{
coords2spdf(points, name)
}
\arguments{
\item{points}{a 2 column matrix of points. The last point should be equal to the first}

\item{name}{a character name for the SPDF}
}
\description{
Function to make a SpatialPolygonsDataFrame from a matrix of points
}

