% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmre.r
\name{fit.mmre}
\alias{fit.mmre}
\title{Function to fit two state Markov model}
\usage{
fit.mmre(data = NULL, parameters, cov.names = "none", decay = FALSE)
}
\arguments{
\item{data}{a \code{mmre} object}

\item{parameters}{a named list of starting values for the model, assumed model is chosen based on the contents of
this list (i.e., to fit a correlated random effect model \code{parameters} must contain the elements \code{u},
\code{log_sig_u} and \code{cov_par})}

\item{cov.names}{a charater string or vector of covariate names (elements in \code{data}) (only for multiple individuals)}

\item{decay}{logical decay formulation to be used or not}
}
\description{
Function to fit two state Markov model
}
