% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.AIC}
\alias{get.AIC}
\title{Function to calculate AIC of mmre fitted model}
\usage{
get.AIC(fit)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}
}
\description{
Function to calculate AIC of mmre fitted model
}

