% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmre.r
\name{get.mmre.data}
\alias{get.mmre.data}
\title{Function to combine argumets and data into a \code{mmre} object for using \code{fit.mmre}}
\usage{
get.mmre.data(data, parameters, cov.names)
}
\arguments{
\item{data}{a \code{mmre} object}

\item{parameters}{a named list of starting values for the model, assumed model is chosen based on the contents of
this list (i.e., to fit a correlated random effect model \code{parameters} must contain the elements \code{u},
\code{log_sig_u} and \code{cov_par})}

\item{cov.names}{a charater string or vector of covariate names (elements in \code{data}) (only for multiple individuals)}
}
\description{
Function to combine argumets and data into a \code{mmre} object for using \code{fit.mmre}
}

