% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.params}
\alias{get.params}
\title{Function to extract either the fixed or random parameters from model}
\usage{
get.params(fit, random)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}

\item{random}{logical if TRUE will return random paramter estimates}
}
\description{
Function to extract either the fixed or random parameters from model
}
