% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.probs}
\alias{get.probs}
\title{Function to get the transition propabilities from a fitted model}
\usage{
get.probs(fit, t)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}

\item{t}{the time index at which to calculate the probability matrix}
}
\description{
Function to get the transition propabilities from a fitted model
}
