% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\docType{methods}
\name{hist.time}
\alias{hist.time}
\title{Function to plot a histogram of dates}
\usage{
hist.time(t)
}
\arguments{
\item{t}{A vector of class POSIXct}
}
\description{
Function to plot a histogram of dates
}
