% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{ll}
\alias{ll}
\title{Function to extract the log likelihood from the fitted models}
\usage{
ll(fit)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}
}
\description{
Function to extract the log likelihood from the fitted models
}

