% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{lr.test}
\alias{lr.test}
\title{Function to do a likelihood ratio test between two models, null hypothesis
that there is no affect of alternative model \code{fit.alt}.}
\usage{
lr.test(fit.full, fit.alt)
}
\arguments{
\item{fit.full}{object of class \code{mmre} returned by \code{fit.mmre} for full model}

\item{fit.alt}{object of class \code{mmre} returned by \code{fit.mmre} for alternative model}
}
\description{
Function to do a likelihood ratio test between two models, null hypothesis
that there is no affect of alternative model \code{fit.alt}.
}
