% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\docType{methods}
\name{map}
\alias{map}
\alias{map,SpatialPolygonsDataFrame,SpatialPolygonsDataFrame-method}
\alias{map,SpatialPolygonsDataFrame,numeric-method}
\alias{map,data.frame,data.frame-method}
\alias{map,data.frame,numeric-method}
\title{Function to location data
Method to map a SpatialPolygonsDataFrame onto google maps}
\usage{
map(x, zoom)

\S4method{map}{SpatialPolygonsDataFrame,numeric}(x, zoom)

\S4method{map}{data.frame,numeric}(x, zoom)
}
\arguments{
\item{x}{A SpatialPolygonsDataFrame with latitude and longitude projection or a data.frame with longitude and latitude columns}

\item{zoom}{numeric 'zoom' of google map}
}
\description{
Function to location data
Method to map a SpatialPolygonsDataFrame onto google maps
}

