% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmre.r
\name{mmre.mod}
\alias{mmre.mod}
\title{Function that choses C++ template to use for model}
\usage{
mmre.mod(mmre.data = NULL)
}
\arguments{
\item{mmre.data}{a \code{mmre} object}
}
\description{
Function that choses C++ template to use for model
}

