% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmre.r
\docType{class}
\name{mmre}
\alias{mmre}
\alias{mmre-class}
\title{Funtions requred to fit two state Markov model with or without individual effects}
\description{
Funtions requred to fit two state Markov model with or without individual effects
}
\details{
Must compile required TMB templates to use
The mmre class
}
\section{Slots}{

\describe{
\item{data}{a data frame which must contain the following named elements ID,
state, time to be used with \code{fit.mmre}}
\item{summary}{a summary of data, automatiacally created when calling \code{fit.mmre},
filled using \code{mmre.summary}. Includes raw transition probabilities.}
\item{fit_data}{a list of the data required to use \code{fit.mmre} (automatically filled)}
\item{continuous}{logical, should a continuous model be used (default)}
\item{fitted}{a character of the chosen model based on arguments given to \code{fit.mmre},
chosen using \code{mmre.mod}}
\item{cov_names}{a charater string or vector of covariate names (elements in \code{data})
(only for multiple individuals)}
\item{parameters}{a named list of starting values for the model, assumed model is chosen based
on the contents of this list (i.e., to fit a correlated random effect model \code{parameters} must contain
the elements \code{u}, \code{log_sig_u} and \code{cov_par})}
\item{fit}{contains the model output}
\item{sdreport}{summary information of the model}
}
}

