% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmre.r
\name{mmre.summary}
\alias{mmre.summary}
\title{Funtion to get summary information from a \code{mmre} object}
\usage{
mmre.summary(data)
}
\arguments{
\item{data}{a \code{mmre} object}
}
\description{
Funtion to get summary information from a \code{mmre} object
}
