% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{onrange}
\alias{onrange}
\title{Function to find if  estimated locations are on or off range}
\usage{
onrange(x, range)
}
\arguments{
\item{x}{a data frame of locations object}

\item{range}{a SpatialPolygonDataFrame of the range}
}
\value{
a logical vector indicating if successive locations were on or off range
}
\description{
Function to find if  estimated locations are on or off range
}

