% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{q2p}
\alias{q2p}
\title{Function to turn off diagonal 2 state transition matrix elements into a transition probability matrix}
\usage{
q2p(q, t)
}
\arguments{
\item{q}{a numeric vector of length two of the off diagonal transition matrix (2 state)}

\item{t}{the time at which to calculate, by default is 1}
}
\description{
Function to turn off diagonal 2 state transition matrix elements into a transition probability matrix
}
