% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{seqMat}
\alias{seqMat}
\title{calculates a sequence matrix from consecutive TRUE/FALSE vector}
\usage{
seqMat(x)
}
\arguments{
\item{x}{a logical vector of consecutive occurances}
}
\value{
a matrix with transition probabilities
}
\description{
calculates a sequence matrix from consecutive TRUE/FALSE vector
}
