% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show.all}
\alias{show.all}
\title{Function to plot all animal tracks in one dataset along with the associated spatial ploygon range. Colours of tracks are created to
be as distinct as possible.}
\usage{
show.all(x, range)
}
\arguments{
\item{x}{a data frame with named elements \code{longitude}, \code{latitude}, and \code{TagID} of animal}

\item{range}{a SpatialPolygonsDataFrame of the associated spatial range}
}
\description{
Function to plot all animal tracks in one dataset along with the associated spatial ploygon range. Colours of tracks are created to
be as distinct as possible.
}

