% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show.duration}
\alias{show.duration}
\title{Function to show duration of exposure data}
\usage{
show.duration(x)
}
\arguments{
\item{x}{a data frame with the following named elements: \code{Start}, start time as a date
object of exposure start times; \code{End}, end time as a date object of exposure end times;
\code{Type}, a factor vector of exposure types.}
}
\description{
Function to show duration of exposure data
}
