% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show.probs}
\alias{show.probs}
\title{Function to plot the (off diagonal) transition probabilties over a specified time period}
\usage{
show.probs(fit, time)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}

\item{time}{a numeric vector of times}
}
\description{
Function to plot the (off diagonal) transition probabilties over a specified time period
}

