% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show.random}
\alias{show.random}
\title{Function to show the individual normal random effects of fitted model retturned by \code{fit.mmre}}
\usage{
show.random(fit)
}
\arguments{
\item{fit}{object of class \code{mmre} returned by \code{fit.mmre}}
}
\description{
Function to show the individual normal random effects of fitted model retturned by \code{fit.mmre}
}

