% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mmre.r
\name{sim.mmre.decay}
\alias{sim.mmre.decay}
\title{Function to simulate Matrkov transition model where transition intensities are
given by beta0 + beta1*exp(-beta2*cov) + individual.random.effect
Number simulated will correspond to length of given times and covariates.}
\usage{
sim.mmre.decay(par.sim, cov, times, ID, start.state, random, fit)
}
\arguments{
\item{par.sim}{a named list of parameters. Must contain "log_baseline","jump", and "decay" all vectors of length two corresponding
to parameters of the off diagonal transition rates. If data being simulated from a random model then must contain element
"sig" the  standard deviation of the bivariate normal individual level random effect.}

\item{cov}{a vector of covariate values (i.e., days since exposure)}

\item{times}{numeric vector of observed times}

\item{ID}{a character vector of whale Ids for zero meaned mvn independent random effects}

\item{start.state}{numeric initial state (can only be 1 or 2)}

\item{random}{logical, should individual random effects be included.}

\item{fit}{a mmre fit object, either supplied or all other armugents}
}
\description{
Function to simulate Matrkov transition model where transition intensities are
given by beta0 + beta1*exp(-beta2*cov) + individual.random.effect
Number simulated will correspond to length of given times and covariates.
}
