% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{sl2sp}
\alias{sl2sp}
\title{Function to change a SpatialLinesDataFrame to a SpatialPolygonsDataFrame}
\usage{
sl2sp(x)
}
\arguments{
\item{x}{An object of class SpatialLinesDataFrame}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Function to change a SpatialLinesDataFrame to a SpatialPolygonsDataFrame
}
\details{
This function is based on code found in the book Applied Spaatial Data
Analysis with R, 2nd edition, Roger S. Bivand, Edzer J. Pebesma and V. Gomez-Rubio
UseR! Series, Springer New York (section 2.6 pp 41)
}
