% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{trans.matrix}
\alias{trans.matrix}
\title{Function to get transition matrix}
\usage{
trans.matrix(x, prob)
}
\arguments{
\item{x}{a matrix of states}

\item{prob}{logical, should probabilities be returned and not transitions.}
}
\description{
Function to get transition matrix
}

