% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mmre.r
\name{sim.mmre.decay}
\alias{sim.mmre.decay}
\title{Function to simulate Matrkov transition model where transition intensities are
given by beta0 + beta1*exp(beta2*cov) + individual.random.effect
Number simulated will correspond to length of given times and covariates.}
\usage{
sim.mmre.decay(par.sim, cov, times, ID, start.state, log_sig_u)
}
\arguments{
\item{par.sim}{a 2 by three matrix of parameters. Each row coresponds to the off diagonal transition intensities.
Each column corresponds to the covariates.}

\item{cov}{a vector of covariate values}

\item{times}{numeric vector of observed times}

\item{ID}{a character vector of whale Ids for zero meaned mvn independent random effects}

\item{start.state}{numeric initial state (can only be 1 or 2)}

\item{log_sig_u}{numeric log sigma for the mvn effect}
}
\description{
Function to simulate Matrkov transition model where transition intensities are
given by beta0 + beta1*exp(beta2*cov) + individual.random.effect
Number simulated will correspond to length of given times and covariates.
}

