% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mmre.r
\name{sim.mmre}
\alias{sim.mmre}
\title{Function to simulate a two state Markov transition baseline model
Number simulated will correspond to length of given times and covariate.}
\usage{
sim.mmre(par.sim, times, ID, start.state, random, fit)
}
\arguments{
\item{par.sim}{a 2 by three matrix of parameters. Each row coresponds to the off diagonal transition intensities.
Each column corresponds to the covariates.}

\item{times}{numeric vector of observed times}

\item{ID}{a character vector of whale Ids for zero meaned mvn independent random effects}

\item{start.state}{numeric initial state (can only be 1 or 2)}

\item{random}{logical, should individual random effects be included.}

\item{fit}{a mmre fit object, either supplied or all other armugents}
}
\description{
Function to simulate a two state Markov transition baseline model
Number simulated will correspond to length of given times and covariate.
}

